+++ Backend/LogicPersistence/LogicPersistence.Api/Models/DTOs/VolunteerDtos.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using System.Numerics;
 
 namespace LogicPersistence.Api.Models.DTOs;
 
@@ -27,7 +26,7 @@ public class VolunteerCreateDto {
 
 	[Required]
 	[Range(100000000, 99999999999)]
-	public BigInteger phone_number { get; set; }
+	public int phone_number { get; set; }
 
 	[Required]
 	[MaxLength(255)]
@@ -67,7 +66,7 @@ public class VolunteerUpdateDto {
 
 	[Required]
 	[Range(100000000, 99999999999)]
-	public BigInteger phone_number { get; set; }
+	public int phone_number { get; set; }
 
 	[Required]
 	[MaxLength(255)]
@@ -87,7 +86,7 @@ public class VolunteerDisplayDto {
 	public string name { get; set; } = string.Empty;
 	public string surname { get; set; } = string.Empty;
 	public int prefix { get; set; }
-	public BigInteger phone_number { get; set; }
+	public int phone_number { get; set; }
 	public string address { get; set; } = string.Empty;
 	public string identification { get; set; } = string.Empty;
